fix: ImageUploader와 PostImageService 통합 (spring-boot-starter-web 의존성을 service layer로 이동)
