+++ api/build.gradle.kts
@@ -1,7 +1,6 @@
 dependencies {
     testImplementation("org.junit.jupiter:junit-jupiter-api:5.7.0")
     testRuntimeOnly ("org.junit.jupiter:junit-jupiter-engine:5.7.0")
-    implementation ("org.springframework.boot:spring-boot-starter-web")
     testImplementation ("io.rest-assured:rest-assured:4.1.2")
     testImplementation ("io.rest-assured:rest-assured-all:4.1.2")
     implementation ("org.springframework.boot:spring-boot-starter-validation")
