+++ api/src/main/java/com/team/post/dto/response/SavePostResponse.java
@@ -5,27 +5,12 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-import java.util.List;
-
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class SavePostResponse {
-    private Long userId;
     private Long postId;
-    private String content;
-    private List<String> postImages;
-    private List<Long> taggedUserIds;
-    private List<String> taggedKeywords;
-    private LocalDateTime createdAt;
 
     public SavePostResponse(SavePostOutput output) {
-        this.userId = output.getUserId();
         this.postId = output.getPostId();
-        this.content = output.getContent();
-        this.postImages = output.getPostImages();
-        this.taggedUserIds = output.getTaggedUserIds();
-        this.taggedKeywords = output.getTaggedKeywords();
-        this.createdAt = output.getCreatedAt();
     }
 }
