+++ service/build.gradle.kts
@@ -9,4 +9,5 @@ tasks.named("jar") {
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter")
     implementation ("org.springframework.boot:spring-boot-starter-mail")
+    api ("org.springframework.boot:spring-boot-starter-web")
 }
