+++ service/src/main/java/com/team/post/dto/input/SavePostInput.java
@@ -4,6 +4,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -12,13 +13,15 @@ import java.util.List;
 public class SavePostInput {
     private Long userId;
     private String content;
+    private List<MultipartFile> postImages;
     private List<Long> taggedUserIds;
     private List<String> taggedKeywords;
 
     @Builder
-    public SavePostInput(Long userId, String content, List<Long> taggedUserIds, List<String> taggedKeywords) {
+    public SavePostInput(Long userId, String content, List<MultipartFile> postImages, List<Long> taggedUserIds, List<String> taggedKeywords) {
         this.userId = userId;
         this.content = content;
+        this.postImages = postImages;
         this.taggedUserIds = taggedUserIds;
         this.taggedKeywords = taggedKeywords;
     }
