+++ service/src/test/java/com/team/post/PostServiceTest.java
@@ -94,7 +94,7 @@ class PostServiceTest {
         List<PostTaggedKeyword> keywordTags = Arrays.asList(keywordTag1, keywordTag2);
 
         given(userService.findUserById(any())).willReturn(user1).willReturn(user2).willReturn(user3);
-        given(postImageService.findImagesByIds(any())).willReturn(postImages);
+        given(postImageService.storeImages(any())).willReturn(postImages);
         given(postRepository.save(any())).willReturn(post);
         given(postTaggedUserService.tagAll(any(), any())).willReturn(userTags);
         given(postTaggedKeywordService.tagAll(any(), any())).willReturn(keywordTags);
@@ -102,21 +102,13 @@ class PostServiceTest {
         SavePostInput input = SavePostInput.builder()
                 .userId(user1.getId())
                 .content("test-content")
-                .postImageIds(postImages.stream().map(PostImage::getId).collect(Collectors.toList()))
                 .taggedUserIds(Arrays.asList(user2.getId(), user3.getId()))
                 .taggedKeywords(Arrays.asList("inMoongram", "spring"))
                 .build();
 
         SavePostOutput output = postService.save(input);
 
-        assertThat(output.getContent()).isEqualTo(input.getContent());
-        assertThat(output.getPostImages().size()).isEqualTo(2);
-        assertThat(output.getTaggedUserIds().size()).isEqualTo(2);
-        assertThat(output.getTaggedUserIds().get(0)).isEqualTo(user2.getId());
-        assertThat(output.getTaggedUserIds().get(1)).isEqualTo(user3.getId());
-        assertThat(output.getTaggedKeywords().size()).isEqualTo(2);
-        assertThat(output.getTaggedKeywords().get(0)).isEqualTo(keywordTag1.getTagKeyword().getKeyword());
-        assertThat(output.getTaggedKeywords().get(1)).isEqualTo(keywordTag2.getTagKeyword().getKeyword());
+        assertThat(output.getPostId()).isEqualTo(post.getId());
     }
 
     @Test
